+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KJAAPMCE46R75WRZRHXD61RQ"
+        "01KJAK6AKZXXCA121SAEPH29G4"
     )
     r"""The unique identifier of the data source"""