+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KJAAPMCJMVF9QG6K13CEDTNT"
+        "01KJAK6AM3SPVJMPX95JGGWNR5"
     )
     r"""The unique identifier of the data source"""