+++ packages/orq-rc/RELEASES.md
@@ -6109,3 +6109,13 @@ Based on:
 - [python v4.5.0-rc.5] packages/orq-rc
 ### Releases
 - [PyPI v4.5.0-rc.5] https://pypi.org/project/orq-ai-sdk/4.5.0-rc.5 - packages/orq-rc
+
+## 2026-02-25 14:30:38
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.729.0 (2.841.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.5.0-rc.6] packages/orq-rc
+### Releases
+- [PyPI v4.5.0-rc.6] https://pypi.org/project/orq-ai-sdk/4.5.0-rc.6 - packages/orq-rc